Remove deduplicating logic.
The neeed for deduplicating was an auxilliary side effect of looking up
MRO wrong with ->can();

When not using ->can(), any subs that appear directly in a package stash
do appear exactly where they seem to appear in the MRO list.

Thus, any exact duplicates of a sub in the heirarchy will in fact **call
themselves** when ascending up the method resolution table.

However, there is presently no useful way for this to work, as a subref
calling itself on a higher inheritance level invokes a perpetual loop
due to next::method assuming the "sub name" to be enough to determine
the calling level.

hence,

   sub Foo::A::meth { $_[0]->next::method }
   *Bar::Baz::meth = \*Foo::A::meth
   Bar::Baz->meth

Will end up with

    Bar::Baz::meth

Assuming it is actually

    Foo::A::meth

And trying to enter the parent of Foo::A ...

So we're going to just keep the subs in place to help you see something
like that could happen if you're careful enough

See $#10 for details.